requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
paramiko==3.4.0
python-dotenv==1.0.1
cryptography==42.0.5
Flask-Bcrypt==1.0.1
